+++ backend/src/filters/stateKeywords.json
@@ -1,0 +1,320 @@
+[
+  "Alabama",
+  "Alabaman",
+  "Alabamian",
+  "Bama",
+  "Heart of Dixie",
+  "Montgomery",
+  "Birmingham",
+  "Mobile",
+  "Alaska",
+  "Alaskan",
+  "Last Frontier",
+  "Anchorage",
+  "Juneau",
+  "Fairbanks",
+  "Arizona",
+  "Arizonan",
+  "Grand Canyon State",
+  "Phoenix",
+  "Tucson",
+  "Mesa",
+  "Arkansas",
+  "Arkansan",
+  "Natural State",
+  "Little Rock",
+  "Fayetteville",
+  "California",
+  "Californian",
+  "Cali",
+  "Golden State",
+  "Los Angeles",
+  "San Francisco",
+  "Hollywood",
+  "Sacramento",
+  "San Diego",
+  "Silicon Valley",
+  "Colorado",
+  "Coloradan",
+  "Centennial State",
+  "Denver",
+  "Boulder",
+  "Rockies",
+  "Connecticut",
+  "Connecticuter",
+  "Nutmegger",
+  "Constitution State",
+  "Hartford",
+  "New Haven",
+  "Delaware",
+  "Delawarean",
+  "First State",
+  "Dover",
+  "Wilmington",
+  "Florida",
+  "Floridian",
+  "Sunshine State",
+  "Miami",
+  "Orlando",
+  "Tampa",
+  "Jacksonville",
+  "Everglades",
+  "Georgia",
+  "Georgian",
+  "Peach State",
+  "Atlanta",
+  "Savannah",
+  "Augusta",
+  "Hawaii",
+  "Hawaiian",
+  "Aloha State",
+  "Honolulu",
+  "Maui",
+  "Oahu",
+  "Kauai",
+  "Big Island",
+  "Idaho",
+  "Idahoan",
+  "Gem State",
+  "Boise",
+  "Twin Falls",
+  "Illinois",
+  "Illinoisan",
+  "Prairie State",
+  "Chicago",
+  "Springfield",
+  "Chi-Town",
+  "Indiana",
+  "Hoosier",
+  "Indianapolis",
+  "Fort Wayne",
+  "Iowa",
+  "Iowan",
+  "Hawkeye State",
+  "Des Moines",
+  "Kansas",
+  "Kansan",
+  "Sunflower State",
+  "Topeka",
+  "Wichita",
+  "Kentucky",
+  "Kentuckian",
+  "Bluegrass State",
+  "Louisville",
+  "Lexington",
+  "Bourbon Country",
+  "Louisiana",
+  "Louisianan",
+  "Pelican State",
+  "New Orleans",
+  "NOLA",
+  "Baton Rouge",
+  "Cajun Country",
+  "Bayou",
+  "Maine",
+  "Mainer",
+  "Pine Tree State",
+  "Portland",
+  "Bangor",
+  "Acadia",
+  "Maryland",
+  "Marylander",
+  "Old Line State",
+  "Baltimore",
+  "Annapolis",
+  "Massachusetts",
+  "Massachusettsan",
+  "Masshole",
+  "Bay State",
+  "Boston",
+  "Cape Cod",
+  "Salem",
+  "Michigan",
+  "Michigander",
+  "Michiganean",
+  "Great Lakes State",
+  "Detroit",
+  "Ann Arbor",
+  "Minnesota",
+  "Minnesotan",
+  "North Star State",
+  "Twin Cities",
+  "Minneapolis",
+  "St. Paul",
+  "Mississippi",
+  "Mississippian",
+  "Magnolia State",
+  "Jackson",
+  "Gulfport",
+  "Missouri",
+  "Missourian",
+  "Show-Me State",
+  "St. Louis",
+  "Kansas City",
+  "Ozarks",
+  "Branson",
+  "Montana",
+  "Montanan",
+  "Big Sky Country",
+  "Helena",
+  "Billings",
+  "Nebraska",
+  "Nebraskan",
+  "Cornhusker State",
+  "Omaha",
+  "Lincoln",
+  "Nevada",
+  "Nevadan",
+  "Silver State",
+  "Las Vegas",
+  "Reno",
+  "Sin City",
+  "New Hampshire",
+  "New Hampshirite",
+  "Granite State",
+  "Concord",
+  "Manchester",
+  "New Jersey",
+  "New Jerseyan",
+  "Garden State",
+  "Jersey",
+  "Newark",
+  "Atlantic City",
+  "New Mexico",
+  "New Mexican",
+  "Land of Enchantment",
+  "Santa Fe",
+  "Albuquerque",
+  "Roswell",
+  "New York",
+  "New Yorker",
+  "Empire State",
+  "New York City",
+  "Brooklyn",
+  "Queens",
+  "Manhattan",
+  "Bronx",
+  "Staten Island",
+  "Buffalo",
+  "Albany",
+  "North Carolina",
+  "North Carolinian",
+  "Tar Heel State",
+  "Charlotte",
+  "Raleigh",
+  "Durham",
+  "Asheville",
+  "North Dakota",
+  "North Dakotan",
+  "Peace Garden State",
+  "Fargo",
+  "Bismarck",
+  "Ohio",
+  "Ohioan",
+  "Buckeye State",
+  "Cleveland",
+  "Columbus",
+  "Cincinnati",
+  "Oklahoma",
+  "Oklahoman",
+  "Sooner State",
+  "Oklahoma City",
+  "Tulsa",
+  "Oregon",
+  "Oregonian",
+  "Beaver State",
+  "Portland",
+  "Eugene",
+  "Salem",
+  "Pennsylvania",
+  "Pennsylvanian",
+  "Keystone State",
+  "Philadelphia",
+  "Philly",
+  "Pittsburgh",
+  "Hershey",
+  "Rhode Island",
+  "Rhode Islander",
+  "Ocean State",
+  "Providence",
+  "Newport",
+  "South Carolina",
+  "South Carolinian",
+  "Palmetto State",
+  "Charleston",
+  "Columbia",
+  "Myrtle Beach",
+  "South Dakota",
+  "South Dakotan",
+  "Mount Rushmore State",
+  "Sioux Falls",
+  "Rapid City",
+  "Black Hills",
+  "Tennessee",
+  "Tennessean",
+  "Volunteer State",
+  "Nashville",
+  "Memphis",
+  "Knoxville",
+  "Chattanooga",
+  "Texas",
+  "Texan",
+  "Lone Star State",
+  "Houston",
+  "Dallas",
+  "Austin",
+  "San Antonio",
+  "Fort Worth",
+  "El Paso",
+  "Utah",
+  "Utahn",
+  "Beehive State",
+  "Salt Lake City",
+  "Provo",
+  "Moab",
+  "Zion",
+  "Vermont",
+  "Vermonter",
+  "Green Mountain State",
+  "Burlington",
+  "Montpelier",
+  "Virginia",
+  "Virginian",
+  "Old Dominion",
+  "Richmond",
+  "Virginia Beach",
+  "Norfolk",
+  "Washington",
+  "Washingtonian",
+  "Evergreen State",
+  "Seattle",
+  "Spokane",
+  "Tacoma",
+  "Olympia",
+  "West Virginia",
+  "West Virginian",
+  "Mountain State",
+  "Charleston",
+  "Morgantown",
+  "Appalachia",
+  "Wisconsin",
+  "Wisconsinite",
+  "Badger State",
+  "Milwaukee",
+  "Madison",
+  "Green Bay",
+  "Wyoming",
+  "Wyomingite",
+  "Equality State",
+  "Cheyenne",
+  "Jackson Hole",
+  "Yellowstone",
+  "New England",
+  "Deep South",
+  "Midwest",
+  "Pacific Northwest",
+  "Appalachian",
+  "Rust Belt",
+  "Bible Belt",
+  "Great Plains"
+]
