+++ backend/src/routes/movie-db.js
@@ -2,6 +2,7 @@ const express = require("express");
 const router = express.Router();
 const { MovieDb } = require("moviedb-promise");
 const moviedb = new MovieDb(process.env.TMDB_API_KEY);
+const keywords = require("../filters/stateKeywords.json");
 
 router.get("/get-media", async (req, res) => {
   try {
@@ -23,11 +24,27 @@ router.get("/get-media", async (req, res) => {
       data = await moviedb.movieInfo({ id });
     }
 
-    return res.json(data);
+    data.overview = sanitize(data.overview);
+
+    const filtered_data = {
+      overview: data.overview,
+      title: data.title,
+      poster_path: data.poster_path,
+    };
+
+    return res.json(filtered_data);
   } catch (error) {
     console.error("Error fetching movie:", error);
     res.status(500).json({ error: "Failed to fetch movie" });
   }
 });
+
+function sanitize(text) {
+  for (const word of keywords) {
+    const regex = new RegExp(`\\b${word}\\b`, "gi");
+    text = text.replace(regex, "_____");
+  }
+  return text;
+}
 
 module.exports = router;
